fine tune the output of the sync cmd
