+++ lib/commands/slicemachine/commands/create-slice/index.js
@@ -2,60 +2,13 @@ import fs from 'fs';
 import path from 'path';
 import inquirer from 'inquirer';
 import consola from 'consola';
-import { promisify } from 'util';
-import cpy from 'copy-template-dir';
-import { hyphenate, pascalize } from 'sm-commons/utils/str';
+import { hyphenate } from 'sm-commons/utils/str';
 
-import Mustache from 'mustache';
-
-import Communication from '../../../../services/communication';
-import Sentry from '../../../../services/sentry';
-import Authentication from '../../../authentication';
 import { ctx } from '../../../../context';
 import getFramework from '../../methods/getFramework';
 import Libraries from '../../common/libraries';
 import Endpoints from '../../misc/endpoints';
-
-const copy = promisify(cpy);
-
-// should move to sm-commons
-function snakecalize(str) {
-  return hyphenate(str).replace(/-/g, '_');
-}
-
-function findRichTextField(model) {
-  const { 'non-repeat': fields = {} } = model;
-  const maybeEntry = Object.entries(fields).find(([, e]) => e.type === 'StructuredText');
-  return maybeEntry ? `slice.primary.${maybeEntry[0]}` : 'false';
-}
-
-function validateSlicezones(customType, selectedSlicezones, availableSlicezones, sliceName) {
-  const errors = selectedSlicezones
-    .map((selectedZone) => {
-      if (availableSlicezones[selectedZone].includes(sliceName)) {
-        return `- The slicezone '${selectedZone}' already contains a slice named ${sliceName}`;
-      }
-      return null;
-    })
-    .filter(a => !!a);
-  if (errors.length) return `[Conflicts] ${customType}: \n ${errors.join('\n')}`;
-  return null;
-}
-
-function validateCustomTypes(selectedCustomTypes, availableCustomTypes, sliceName) {
-  const errors = selectedCustomTypes
-    .map((selectedCT) => {
-      const slicezones = availableCustomTypes[selectedCT];
-      const szoneIds = Object.keys(slicezones);
-      // we only make this validation if there is only one slicezone otherwise we wait for the user to select the slicezones he wants to put his slice in.
-      if (szoneIds.length === 1) return validateSlicezones(selectedCT, szoneIds, slicezones, sliceName);
-      return null;
-    })
-    .filter(a => !!a);
-
-  if (errors.length) return errors.join('\n');
-  return null;
-}
+import Slices from '../../common/slices';
 
 const Prompts = {
   componentName(pathToCreate) {
@@ -77,102 +30,14 @@ const Prompts = {
       },
     }]);
   },
-
-  customTypes(customTypeList, sliceName) {
-    const customTypesIds = Object.keys(customTypeList);
-    return inquirer.prompt([{
-      type: 'checkbox',
-      message: 'Select custom types',
-      name: 'customtypes',
-      choices: customTypesIds.map(ctId => ({ name: ctId })),
-      validate(answers) {
-        if (answers.length < 1) return 'You must choose at least one custom type.';
-
-        const errors = validateCustomTypes(answers, customTypeList, sliceName);
-        if (errors) return errors;
-
-        return true;
-      },
-    }]);
-  },
-
-  slicezones(customTypeId, slicezoneList, sliceName) {
-    return inquirer.prompt([{
-      type: 'checkbox',
-      message: `Select slicezones for ${customTypeId}`,
-      name: customTypeId,
-      choices: Object.keys(slicezoneList).map(sz => ({ name: sz })),
-      validate(answers) {
-        if (answers.length < 1) return 'You must choose at least one slice zone.';
-
-        const errors = validateSlicezones(customTypeId, answers, slicezoneList, sliceName);
-        if (errors) return errors;
-
-        return true;
-      },
-    }]);
-  },
 };
-
-async function createRemoteSlice(framework, prismicEndpoint, sliceName) {
-  try {
-    await Authentication.connect();
-
-    const customTypes = await Communication.getAsJson(Endpoints.Prismic.customTypes(prismicEndpoint), { withSlices: true });
-
-    const slicezoneByCustomTypes = customTypes.reduce((acc, ct) => {
-      const slicesBySlicezone = Object.entries(ct.slices).reduce((slicesAcc, [szKey, slices]) => ({ ...slicesAcc, [szKey]: Object.keys(slices).map(pascalize) }), {});
-      return { ...acc, [ct.mask_id]: slicesBySlicezone };
-    }, {});
-    const { customtypes: selectedCustomTypes } = await Prompts.customTypes(slicezoneByCustomTypes, sliceName);
-    const insertOptions = await (async () => {
-      let acc = {};
-      for (let i = 0; i < selectedCustomTypes.length; i += 1) {
-        const ctId = selectedCustomTypes[i];
-        const slicezones = slicezoneByCustomTypes[ctId];
-        const szonesIds = Object.keys(slicezones);
-        if (szonesIds.length <= 1) acc = { ...acc, [ctId]: szonesIds };
-        else {
-          /* eslint-disable no-await-in-loop */
-          const answers = await Prompts.slicezones(ctId, slicezones, sliceName);
-          acc = { ...acc, [ctId]: answers[ctId] };
-        }
-      }
-      return acc;
-    })();
-
-
-    const sliceModel = (() => {
-      const defaultModelFile = fs.readFileSync(path.join(__dirname, 'templates', framework, 'model.json'), 'utf8');
-      return JSON.parse(Mustache.render(defaultModelFile, { componentName: sliceName }));
-    })();
-
-    const sliceId = snakecalize(sliceName);
-
-    const opts = {
-      id: sliceId,
-      model: sliceModel,
-      customtypes: insertOptions,
-    };
-    await Communication.postJson(`${Endpoints.Prismic.slices(prismicEndpoint)}/create`, opts);
-    return { model: sliceModel };
-  } catch (e) {
-    Sentry.report(e);
-    consola.info(`Could not create slice ${sliceName} in Prismic.`);
-    throw new Error(`${e}`);
-  }
-  // - make the request to prismic and if it fails, display the errors and how to resolve it otherwise run the copy below =>
-}
 
 /* eslint-disable  consistent-return */
 export default async function () {
   const framework = await getFramework(ctx.SliceMachine);
-  if (!framework || framework instanceof Error === true) {
-    return consola.info('Exiting...');
-  }
-  const libPath = await Libraries.selectLocal();
-  const { pathToSlices } = Libraries.infos(libPath);
-  const { sliceName } = await Prompts.componentName(pathToSlices);
+  const lib = await Libraries.selectLocal();
+  const { pathToSlices: libPath } = Libraries.infos(lib);
+  const { sliceName } = await Prompts.componentName(libPath);
 
   const prismicEndpoint = Endpoints.Prismic.fromCtxOrConfig();
   if (!prismicEndpoint) {
@@ -180,18 +45,8 @@ export default async function () {
   }
 
   try {
-    const { model } = await createRemoteSlice(framework, prismicEndpoint, sliceName);
-
-    const field = findRichTextField(model);
-    await copy(
-      path.join(__dirname, 'templates', framework),
-      path.join(pathToSlices, sliceName),
-      {
-        componentName: sliceName,
-        colorCode: '#FFFF63',
-        field,
-      },
-    );
+    await Slices.createRemote(framework, prismicEndpoint, sliceName);
+    await Slices.createLocal(framework, sliceName, /* modelWithPath */ null, libPath, path.join(__dirname, '../../templates'));
   } catch (e) {
     return consola.error(e.message);
   }
