+++ lib/commands/slicemachine/commands/init.js
@@ -86,101 +86,101 @@ function handleRecap(manifest, library, info) {
 }
 
 async function init() {
-  const {
-    yes,
-  } = ctx.SliceMachine;
+  try {
+    const {
+      yes,
+    } = ctx.SliceMachine;
 
-  if (yes) {
-    consola.info('Running in no-confirm mode: some files might be overridden.');
-  }
-
-  const framework = await getFramework(ctx.SliceMachine);
-
-  if (!framework || framework instanceof Error === true) {
-    return consola.info('Exiting...');
-  }
-
-  const {
-    zip, manifest, library,
-  } = await (async function handleZip() {
-    const spinner = ora('Downloading files from server').start();
-    const endpoint = urljoin(ctx.SliceMachine.apiEndpoint, 'bootstrap');
-    const maybeZipPath = await SliceMachine.downloadSlices(endpoint, {
-      framework,
-      lib: ctx.SliceMachine.lib,
-    });
-    spinner.succeed();
-    if (!maybeZipPath || maybeZipPath instanceof Error === true) {
-      consola.error('An error occured while downloading files from server.');
-      return { zip: new Error() };
+    if (yes) {
+      consola.info('Running in no-confirm mode: some files might be overridden.');
     }
-    const z = (() => {
-      try {
-        return new AdmZip(maybeZipPath);
-      } catch (e) {
-        consola.error(`Unexpected error: could not instantiate Zip.\n[Full error] ${e}`);
-        return new Error();
+
+    const framework = await getFramework(ctx.SliceMachine);
+
+    const {
+      zip, manifest, library,
+    } = await (async function handleZip() {
+      const spinner = ora('Downloading files from server').start();
+      const endpoint = urljoin(ctx.SliceMachine.apiEndpoint, 'bootstrap');
+      const maybeZipPath = await SliceMachine.downloadSlices(endpoint, {
+        framework,
+        lib: ctx.SliceMachine.lib,
+      });
+      spinner.succeed();
+      if (!maybeZipPath || maybeZipPath instanceof Error === true) {
+        consola.error('An error occured while downloading files from server.');
+        return { zip: new Error() };
       }
+      const z = (() => {
+        try {
+          return new AdmZip(maybeZipPath);
+        } catch (e) {
+          consola.error(`Unexpected error: could not instantiate Zip.\n[Full error] ${e}`);
+          return new Error();
+        }
+      })();
+      if (z instanceof Error === true) {
+        return z;
+      }
+      const bootFile = JSON.parse(z.readAsText('boot.json'));
+      z.deleteFile('boot.json');
+      return {
+        zip: z,
+        ...bootFile,
+      };
+    }());
+
+    if (!zip || zip instanceof Error === true) {
+      return consola.info('Exiting...');
+    }
+
+    const [error, info] = await (async () => {
+      const _format = (prismicDomain, magicUrl, apiEndpoint = ctx.endpoint) => Object.assign(
+        { prismicDomain },
+        magicUrl && { magicUrl },
+        apiEndpoint && { apiEndpoint },
+      );
+
+      if (!ctx.SliceMachine.noPrismic) {
+        const ct = JSON.parse(zip.readAsText('mergedCustomTypes.json'));
+        consola.info('Next: create a fresh Prismic repository');
+        const r = await createPrismicRepo(ct);
+
+        const [protocol, base] = ctx.base.split('://');
+        if (!protocol || !base) return [`Base url "${ctx.base}" is invalid: did you forget to specify protocol?`, null];
+
+        const apiEndpoint = `${protocol}://${r.domain}.${protocol === 'https' ? 'cdn.' : ''}${ctx.base.replace(/^(?:https?:\/\/)?(?:www\.)?/i, '')}/api/v2`;
+        return [null, _format(r.domain, r.magicUrl, apiEndpoint)];
+      }
+      return [null, _format()];
     })();
-    if (z instanceof Error === true) {
-      return z;
-    }
-    const bootFile = JSON.parse(z.readAsText('boot.json'));
-    z.deleteFile('boot.json');
-    return {
-      zip: z,
-      ...bootFile,
-    };
-  }());
 
-  if (!zip || zip instanceof Error === true) {
-    return consola.info('Exiting...');
+    if (error) return consola.error(error);
+
+    zip.deleteFile('mergedCustomTypes.json');
+
+    const spinner = ora('Installing dependencies\n').start();
+    const packageManager =
+      existsSync(path.join(process.cwd(), 'yarn.lock'))
+        ? PackageManager.yarn
+        : PackageManager.npm;
+    handleDependencies(manifest, packageManager);
+    spinner.stop();
+
+    (function handleSmFile() {
+      const sm = {
+        libraries: [ctx.SliceMachine.localSlicesPath, ...manifest.libraries],
+        ...(info.apiEndpoint ? {
+          apiEndpoint: info.apiEndpoint,
+        } : null),
+      };
+      writeFileSync('sm.json', JSON.stringify(sm), 'utf8');
+    }());
+
+    return handleRecap(manifest, library, info);
+  } catch (e) {
+    return consola.error(e);
   }
-
-  const [error, info] = await (async () => {
-    const _format = (prismicDomain, magicUrl, apiEndpoint = ctx.endpoint) => Object.assign(
-      { prismicDomain },
-      magicUrl && { magicUrl },
-      apiEndpoint && { apiEndpoint },
-    );
-
-    if (!ctx.SliceMachine.noPrismic) {
-      const ct = JSON.parse(zip.readAsText('mergedCustomTypes.json'));
-      consola.info('Next: create a fresh Prismic repository');
-      const r = await createPrismicRepo(ct);
-
-      const [protocol, base] = ctx.base.split('://');
-      if (!protocol || !base) return [`Base url "${ctx.base}" is invalid: did you forget to specify protocol?`, null];
-
-      const apiEndpoint = `${protocol}://${r.domain}.${protocol === 'https' ? 'cdn.' : ''}${ctx.base.replace(/^(?:https?:\/\/)?(?:www\.)?/i, '')}/api/v2`;
-      return [null, _format(r.domain, r.magicUrl, apiEndpoint)];
-    }
-    return [null, _format()];
-  })();
-
-  if (error) return consola.error(error);
-
-  zip.deleteFile('mergedCustomTypes.json');
-
-  const spinner = ora('Installing dependencies\n').start();
-  const packageManager =
-    existsSync(path.join(process.cwd(), 'yarn.lock'))
-      ? PackageManager.yarn
-      : PackageManager.npm;
-  handleDependencies(manifest, packageManager);
-  spinner.stop();
-
-  (function handleSmFile() {
-    const sm = {
-      libraries: [ctx.SliceMachine.localSlicesPath, ...manifest.libraries],
-      ...(info.apiEndpoint ? {
-        apiEndpoint: info.apiEndpoint,
-      } : null),
-    };
-    writeFileSync('sm.json', JSON.stringify(sm), 'utf8');
-  }());
-
-  return handleRecap(manifest, library, info);
 }
 
 export default init;
