+++ lib/commands/slicemachine/commands/ls.js
@@ -18,11 +18,11 @@ function displayLibs(libs) {
 async function ls() {
   try {
     const sm = getSmFile();
-    const expandedLibs = await Promise.all(sm.libraries.map(async (libPath) => {
+    const expandedLibs = await Promise.all(sm.libraries.map(async (lib) => {
       try {
-        const slices = await Libraries.slicesModels(libPath);
+        const slices = await Libraries.slicesModels(lib);
         const sliceNames = Object.keys(slices);
-        return [libPath, sliceNames];
+        return [lib, sliceNames];
       } catch (e) {
         console.error(e);
       }
