+++ lib/commands/slicemachine/commands/sync.js
@@ -1,6 +1,9 @@
 import consola from 'consola';
 import ora from 'ora';
 import inquirer from 'inquirer';
+import path from 'path';
+
+import { ctx } from '../../../context';
 
 import Communication from '../../../services/communication';
 import { Files, Folders, Objects } from '../utils';
@@ -10,8 +13,11 @@ import Endpoints from '../misc/endpoints';
 import Sentry from '../../../services/sentry';
 
 import { getOrFail as getSmFile } from '../methods/sm';
+import getFramework from '../methods/getFramework';
+
 import Libraries from '../common/libraries';
 import Slice from '../models/slice';
+import Slices from '../common/slices';
 
 // writes the custom types in the filesystem
 async function writeCustomTypes(customTypes) {
@@ -41,7 +47,7 @@ export default async function () {
 
     const sm = getSmFile();
 
-    const modelsByLib = await Promise.all(sm.libraries.map(libPath => Libraries.slicesModels(libPath)));
+    const modelsByLib = await Promise.all(sm.libraries.map(lib => Libraries.slicesModels(lib)));
 
     // If there are name collision, the first libraries in sm.json are ordered by priority
     // Doing the following will merge all slices together by name with the right priority
@@ -50,8 +56,9 @@ export default async function () {
 
     Object.entries(slicezoneByCustomTypes).forEach(([customTypeId, slicezones]) => {
       Object.entries(slicezones).forEach(async ([slicezoneId, slices]) => {
+        consola.info(`=== Pulling slices for custom type [${customTypeId}] - slicezone [${slicezoneId}] ===\n`);
+
         const [miss, diff, equals] = Objects.matchIn(slices, modelsById);
-        // const spinner = ora(`Pulling slices for custom type [${customTypeId}] - slicezone [${slicezoneId}]\n`).start();
 
         // ALREADY UP TO DATE
         const equalsWithMeta = Object.entries(equals).reduce((acc, [sliceId]) => ({ ...acc, ...Objects.find(sliceId, mergedModelsWithMeta)((searchKey, [, targetValue]) => searchKey === targetValue.id) }), {});
@@ -68,21 +75,11 @@ export default async function () {
           if (!newModel) throw new Error('Oops! Something went wrong while resolving conflicts.');
 
           const newModelWithMeta = Object.entries(oldModelWithMeta).reduce((acc, [sliceName, value]) => ({ ...acc, [sliceName]: { ...value, model: newModel } }), {});
-          console.log(newModelWithMeta);
           return { ...acc, ...newModelWithMeta };
         }, {});
         Displays.handleConflictedSlices(conflictedWithMeta);
-
-        // spinner.succeed();
-
-
-        // these slices are up to date => equals
-        // these slices are missing in your project, select which one you want to pull
-        // these slices are conflicting => diff
       });
     });
-
-    // function matchObjIn(toMatchObj, dataset): [miss, diff, equals]
   } catch (e) {
     Sentry.report(e);
     consola.error(e);
@@ -93,22 +90,33 @@ export default async function () {
 const Displays = {
   equalSlices(equals) {
     const sliceList = Object.keys(equals).map(k => `\t- ${k}`).join('\n');
-    consola.success(`These slices are already in your project: \n${sliceList}`);
+    consola.success(`These slices are already in your project: \n${sliceList}\n`);
   },
 
   async createMissingSlices(missing) {
     if (!missing) return;
+    const sliceList = Object.keys(missing).map(k => `\t- ${k}`).join('\n');
+    consola.info(`These slices doesn't exist in your project: \n${sliceList}\n`);
+
     const { selectedSlices } = await Prompts.createMissingSlices(missing);
 
-    if (selectedSlices.length === 0) returnl;
+    if (selectedSlices.length === 0) return;
 
-    const libPath = await Libraries.selectLocal();
-    const missingWithmeta = Objects.map(missing, ([id, model]) => {
+    const framework = await getFramework(ctx.SliceMachine);
+    const lib = await Libraries.selectLocal();
+    const { pathToSlices: libPath } = await Libraries.infos(lib);
+
+    const selectedMissing = Objects.filter(missing, ([id]) => selectedSlices.includes(id));
+    const missingWithMeta = Objects.map(selectedMissing, ([id, model]) => {
       const name = Slice.nameFromId(id);
-      return Slice.apply(id, name, model, /* isLocal */ true, libPath);
+      return Slice.apply(id, name, model, /* isLocal */ true, lib, libPath);
     });
 
-    console.log(missingWithmeta);
+    const missingEntries = Object.entries(missingWithMeta);
+    for (let i = 0; i < missingEntries.length; i += 1) {
+      const [sliceName, missingSlice] = missingEntries[i];
+      await Slices.createLocal(framework, sliceName, { path: missingSlice.path, model: missingSlice.model }, libPath, path.join(__dirname, '../templates'));
+    }
   },
 
   async handleConflictedSlices(conflicted) {
@@ -128,12 +136,8 @@ const Displays = {
       const modelWithMeta = Objects.find(key, localConflicts)()[key];
       if (modelWithMeta) {
         try {
-          console.log(modelWithMeta);
-          console.log(modelWithMeta.path);
-          console.log(modelWithMeta.model);
           Files.writeJson(modelWithMeta.path, modelWithMeta.model);
         } catch (e) {
-          console.log(e);
           consola.error(`Failed to write the slice ${key} in ${modelWithMeta.path}`);
         }
       } else consola.error(`Unable to update the model of ${key}`);
@@ -170,19 +174,3 @@ const Prompts = {
     }]);
   },
 };
-
-
-// function createSlice(pathToSlices, sliceName) {
-//   try {
-//     await copy(
-//       path.join(__dirname, 'templates', framework),
-//       path.join(pathToSlices, sliceName),
-//       {
-//         componentName: sliceName,
-//         colorCode: '#FFFF63',
-//         field: 'false',
-//       },
-//     );
-//   } catch (e) {
-//     return consola.error(e.message);
-//   }
