+++ lib/commands/slicemachine/common/frameworks.js
@@ -19,8 +19,7 @@ export const determineFrameworkUsed = async (DEBUG = false) => {
   const res = await fetchFrameworks();
   spinner.succeed();
   if (res.status !== 200) {
-    consola.error(`[getFramework/fetch] Unable to fetch frameworks. Error code: ${res.status}`);
-    return [new Error(), null];
+    throw new Error(`[getFramework/fetch] Unable to fetch frameworks. Error code: ${res.status}`);
   }
   const frameworks = await res.json();
   const testsByFramework = frameworks
@@ -33,12 +32,10 @@ export const determineFrameworkUsed = async (DEBUG = false) => {
     .reduce((acc, [currentFmwk, tests]) => acc || (testProject(tests) ? currentFmwk : null), null);
 
   if (!frmwk && !DEBUG) {
-    consola.error('SliceMachine should be launched inside a Next/Nuxt app');
-    consola.info('Did you run this command from the right folder?');
-    return [new Error(), null];
+    throw new Error('SliceMachine should be launched inside a Next/Nuxt app.\nDid you run this command from the right folder?');
   }
   if (DEBUG) {
-    return [null, 'nuxt'];
+    return 'nuxt';
   }
-  return [null, frmwk];
+  return frmwk;
 };
