+++ lib/commands/slicemachine/common/libraries.js
@@ -43,20 +43,20 @@ function infos(libPath) {
   };
 }
 
-function buildSliceModel(libPath, isLocal, slicePath) {
+function buildSliceModel(lib, libPath, isLocal, slicePath) {
   if (!slicePath || slicePath.length === 0) return null;
-  return Slice.fromModelPath(slicePath, isLocal, libPath);
+  return Slice.fromModelPath(slicePath, isLocal, lib, libPath);
 }
 
-async function slicesModels(libPath) {
+async function slicesModels(lib) {
   try {
-    const { pathToSlices, isLocal } = infos(libPath);
+    const { pathToSlices, isLocal } = infos(lib);
 
     const models = await (async () => {
       const paths = await globby([
         `${pathToSlices}/**/model.json`,
       ]);
-      return paths.reduce((acc, p) => ({ ...acc, ...buildSliceModel(libPath, isLocal, p) }), {});
+      return paths.reduce((acc, p) => ({ ...acc, ...buildSliceModel(lib, /* libPath */ pathToSlices, isLocal, p) }), {});
     })();
 
     return models;
