+++ lib/commands/slicemachine/models/slice.js
@@ -2,13 +2,14 @@ import path from 'path';
 import { snakelize, pascalize } from 'sm-commons/utils/str';
 
 export default {
-  apply(id, name, model, isLocal, libPath) {
+  apply(id, name, model, isLocal, lib, libPath) {
     if (!id || !name) return null;
 
     return {
       [name]: {
         id,
-        lib: libPath,
+        lib,
+        libPath: path.normalize(libPath),
         path: path.normalize(path.join(libPath, name, './model.json')),
         model,
         isLocal,
@@ -16,7 +17,7 @@ export default {
     };
   },
 
-  fromModelPath(modelPath, isLocal, libPath) {
+  fromModelPath(modelPath, isLocal, lib, libPath) {
     const name = modelPath
       .replace('/model.json', '')
       .split('/')
@@ -25,7 +26,7 @@ export default {
     const id = snakelize(name);
     const model = require(modelPath);
 
-    return this.apply(id, name, model, isLocal, libPath);
+    return this.apply(id, name, model, isLocal, lib, libPath);
   },
 
   nameFromId(id) {
