+++ lib/commands/slicemachine/templates/nuxt/index.vue
