+++ lib/commands/slicemachine/templates/nuxt/model.json
