+++ lib/commands/slicemachine/utils/files.js
@@ -24,4 +24,8 @@ export default {
   append(filePath, data) {
     fs.appendFileSync(filePath, data);
   },
+
+  copy(src, dest) {
+    fs.copyFileSync(src, dest);
+  },
 };
