+++ lib/commands/slicemachine/utils/objects.js
@@ -1,10 +1,10 @@
 import deepEqual from 'deep-equal';
 
-export default {
-  _defaults: {
-    find: (key, [id]) => id === key,
-  },
+const _defaults = {
+  find: (key, [id]) => id === key,
+};
 
+export default {
   matchIn(toMatchObj, dataset) {
     if (!toMatchObj) return [];
 
@@ -21,7 +21,7 @@ export default {
   find(key, objcts) {
     if (!key || !objcts) return {};
 
-    return function exec(customFind = this._defaults.find) {
+    return function exec(customFind = _defaults.find) {
       const found = Object.entries(objcts).find(([id, value]) => customFind(key, [id, value]));
       if (!found) return {};
